Renamed main directories
