+++ .gitignore
@@ -243,6 +243,6 @@ _Pvt_Extensions
 *.mid
 *.bat
 *.zip
-/VCProject/MEMZ.VC.db
+*.db
 build/
 Build/
