+++ NyanMBR/Source/Stage2/Animation/Image/drawIntroFrame.asm
