+++ NyanMBR/Source/Stage2/Animation/Image/drawNormalFrame.asm
