+++ NyanMBR/Source/Stage2/Animation/Image/initDrawing.asm
