+++ NyanMBR/Source/Stage2/Utils/macros.asm
