+++ WindowsTrojan/MEMZ.sln
