+++ WindowsTrojan/MEMZ/MEMZ.vcxproj
