+++ WindowsTrojan/MEMZ/MEMZ.vcxproj.filters
