+++ WindowsTrojan/MEMZ/win32_crt_float.cpp
